both editors use suggestion list, remove tippy
